refactor: improving header component
